Add neovim (nvim) and neovim settings
This commit updates Brewfile to install neovim and neovim-remote.

Currently, the neovim formula is in a state where it fails to install with homebrew, so we are taking a workaround with the --args HEAD option. Please remove this option once the build is stable.

.config/neovim/init.vim is a configuration file for neovim, but it also works with vim.

If you use vim, execute the following command.

$ ln -snf ${XDG_CONFIG_HOME}/nvim/init.vim ${HOME}/.vimrc