Add missing dependencies: GNU make
This commit enhances Brewfile to install GNU make.

Some environments, e.g. macOS, may use BSD make, and while we are careful to ensure that the dotfiles Makefile works with BSD make, we do not like to continue to use BSD make in terms of environment reproducibility.

Installing GNU make from homebrew will improve the reproducibility of the environment.