Fix conditional expression to enable vim alias
We had defined an alias that replaced vim with nvim only if neovim was installed.
However, it actually used 'nvr' instead of 'nvim', and even if 'nvim' was already installed, 'nvr' was not installed and the vim alias call could fail. When defining vim alias, the existence of nvr as well as nvim is now checked.